refactor(UI): organize search components by layout
