+++ frontend/components/commons/results/RecordExtraActions.vue
