+++ frontend/components/text-classifier/results/ClassifierAnnotationArea.vue
