+++ frontend/components/text-classifier/results/LabelPill.vue
