+++ frontend/components/text-classifier/results/RecordExplain.vue
