+++ frontend/components/text-classifier/results/RecordInputs.vue
