+++ frontend/components/text-classifier/results/RecordTextClassification.vue
