+++ frontend/components/text-classifier/sidebar/TextClassificationProgress.vue
