+++ frontend/components/token-classifier/results/EntitiesHeader.vue
