+++ frontend/components/token-classifier/results/RecordTokenClassification.vue
@@ -1,28 +1,28 @@
 <template>
   <div class="record">
-      <div ref="list" v-if="textSpans.length">
-        <TextSpan
-          v-for="(token, i) in textSpans"
-          :key="i"
-          :record="record"
-          :span-id="i"
-          :spans="textSpans"
-          :dataset="dataset"
-          :class="isSelected(i) ? 'selected' : ''"
-          @startSelection="onStartSelection"
-          @endSelection="onEndSelection"
-          @selectEntity="onSelectEntity"
-          @changeEntityLabel="onChangeEntityLabel"
-          @removeEntity="onRemoveEntity"
-          @reset="onReset"
-        />
-      </div>
-      <RecordExtraActions
-        :allow-change-status="annotationEnabled"
+    <div v-if="textSpans.length" ref="list">
+      <TextSpan
+        v-for="(token, i) in textSpans"
+        :key="i"
         :record="record"
-        @onChangeRecordStatus="onChangeRecordStatus"
-        @onShowMetadata="$emit('onShowMetadata')"
+        :span-id="i"
+        :spans="textSpans"
+        :dataset="dataset"
+        :class="isSelected(i) ? 'selected' : ''"
+        @startSelection="onStartSelection"
+        @endSelection="onEndSelection"
+        @selectEntity="onSelectEntity"
+        @changeEntityLabel="onChangeEntityLabel"
+        @removeEntity="onRemoveEntity"
+        @reset="onReset"
       />
+    </div>
+    <RecordExtraActions
+      :allow-change-status="annotationEnabled"
+      :record="record"
+      @onChangeRecordStatus="onChangeRecordStatus"
+      @onShowMetadata="$emit('onShowMetadata')"
+    />
   </div>
 </template>
