+++ frontend/components/token-classifier/results/TextSpan.vue
