+++ frontend/components/token-classifier/sidebar/SidebarCollapsableMentions.vue
