+++ frontend/components/token-classifier/sidebar/TokenClassificationProgress.vue
