+++ frontend/components/token-classifier/sidebar/TokenClassificationStats.vue
